requirements.txt
Django==4.1.1
Pillow==9.2.0
django-tables2==2.4.1
django-crispy-forms==1.14.0
gunicorn==20.1.0
python-dotenv==0.21.0
pyyaml
jinja2
whitenoise==6.4.0
psycopg[binary]==3.1.8
environs[django]==9.5.0
langchain_openai